+++ include/crpropa/module/Acceleration.h
@@ -168,7 +168,7 @@ class QuasiLinearTheory : public StepLengthModifier {
 /// Thanks to Matthew Weiss, Penn State University for the first work on this
 /// feature in 2017.
 class ParticleSplitting : public Module {
-	int numSplits;
+	int numberSplits;
 	int crossingThreshold;
 	double minWeight;
 	ref_ptr<Surface> surface;
@@ -186,7 +186,7 @@ class ParticleSplitting : public Module {
 	                            multiple splitting modules are present.
 	*/
 	ParticleSplitting(Surface *surface, int crossingThreshold = 50,
-	                  int numSplits = 5, double minWeight = 0.01,
+	                  int numberSplits = 5, double minWeight = 0.01,
 	                  std::string counterid = "ParticleSplittingCounter");
 
 	// update the candidate